+++ CHANGELOG.md
@@ -4,3 +4,4 @@ This document explains the new/changed features in unicef/molo-iogt.
 ## [2.9.0]
 - Version 2.9 will be the final version to support Python 2
 - Added sentry-sdk for application monitoring
+- Added new Wagtail content-type `ExtendedArticlePage` for schema markup